pyproject.toml
[tool.poetry]
name = "airflow-playground"
version = "0.1.0"
description = ""
authors = ["Pritish Jadhav"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.6"
kubernetes = "*"
virtualenv = "*"
pendulum = "<3"
apache-airflow = "==2.9.1"
pandas = "*"
numpy = "*"

[tool.poetry.dev-dependencies]
black = "*"
pytest = "*"
isort = "*"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
